+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Change setup dialog to use new design
 - Change SMB mount error to be more descriptive
 - Change history log to include archive mounting errors
+- Change behavior on missing keyring daemon: Save passwords in memory until close
 - Add docker to containers and VMs exclusion preset
 - Add per-backup preferences window
 - Add ability to set custom titles for backup configs